#27 refactored basic services
