+++ app/Services/Basic/BasicDeleter.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Services\Basic\Deleter;
+namespace App\Services\Basic;
 
 use Illuminate\Database\Eloquent\Model;
+use App\Services\Basic\Contracts\BasicDeleter as Deleter;
 
-class BasicDeleter implements BasicDeleterInterface
+class BasicDeleter implements Deleter
 {
     public function delete(Model $model): void
     {
