+++ app/Services/Basic/BasicUpdater.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Services\Basic\Updater;
+namespace App\Services\Basic;
 
+use App\Services\Basic\Contracts\BasicUpdater as Updater;
 use Illuminate\Database\Eloquent\Model;
 
-class BasicUpdater implements BasicUpdaterInterface
+class BasicUpdater implements Updater
 {
     public function update(Model $model, array $data): Model
     {
