+++ app/Services/Basic/Contracts/BasicUpdater.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Services\Basic\Updater;
+namespace App\Services\Basic\Contracts;
 
 use Illuminate\Database\Eloquent\Model;
 
-interface BasicUpdaterInterface
+interface BasicUpdater
 {
     public function update(Model $model, array $data): Model;
 }
